Thank you for your service, Microsoft Windows *yeet*
It's a shame what with all the portability work I did, but I no longer
have any way of testing this port, and the last Windows version on which
it was *ever* tested -- Vista -- went EOL over a decade ago.

Tests for <net/if_dl.h>, which I think is a Macintosh thing, also
removed for very similar reasons. Same with the last (I think) vestige
of Objective-C++.